selfeval_requirements.txt
# Navigate to this directory and run -- Install via $ pip install -r selfeval_requirements.txt

python-docx==1.1.0
questionary==2.0.1
